Fixes issue with today column
Today column was triggering socket:task.update, and
when recieved wrong info - updated the task
Also today task was breaking js in several places, so no execution was done.

-----------------------------------
Fixes issue with socket task:update

Socket update was broken due to several bugs.

-----------------------------------
Fixes issue with animationSpeed for long distances

Animation for long distances was way too fast, and not clear.
So animation duration now depends on distance between positions.

-----------------------------------
Fixes issue with position order on task movement

Tasks was prepended when moved to a new column or index was 0.
Now prepends tasks only when index was 0,
and appends if task is new for a column.